fix: correct Reference#plain? predicate for structured citations
The plain? method used (source && id && version).nil? which evaluates
to true when version is nil even if source and id are present. This
caused broader/narrower relations and other structured citations to
serialize as empty ref: instead of ref: {source:, id:}.

Fix: source.nil? && id.nil? && version.nil? — any one present means
structured mode.

Also extract Reference class from Citation into its own file, with
Citation as a backward-compatible alias.
